Tratando o erro 404

+++ src/main/java/br/com/alura/forum/TopicosController.java
@@ -3,6 +3,7 @@ package br.com.alura.forum;
 import java.net.URI;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -64,24 +65,36 @@ public class TopicosController {
 	}
 	
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDto detalhar(@PathVariable Long id){
-		Topico topico = topicoRepository.getOne(id);
-		return new DetalhesDoTopicoDto(topico);
+	public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id){
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if(topico.isPresent()){
+			return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get())); 
+		}
+		return ResponseEntity.notFound().build();
 	}
 	
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
-		Topico topico = form.atualizar(id, topicoRepository);
-		
-		return ResponseEntity.ok(new TopicoDto(topico));
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()){
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
+		return ResponseEntity.notFound().build();
 	}
 	
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id){
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build(); 
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()){
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build(); 
+		}
+		return ResponseEntity.notFound().build();
+		
+		
 	}
